+++ src/syscall/fs/handles.c
@@ -4,14 +4,14 @@
 #include "syscall/fs/syscall_fs.h"
 #include "vfs/vfs.h"
 
-ssize_t syscall_routine_open(const char *restrict path, syscall_open_flags_t flags, mode_t mode, proc_t *proc) {
+int syscall_routine_open(const char *restrict path, syscall_open_flags_t flags, mode_t mode, proc_t *proc) {
     printf("Open syscall called with path %s and flags %o.\n", path, flags);
 
-    ssize_t res = 0;
+    int res = 0;
 
     char path_abs[PATH_MAX] = { 0 };
 
-    res = (ssize_t)handle_relative_path(path, proc, path_abs);
+    res = handle_relative_path(path, proc, path_abs);
     if (res < 0) {
         printf("syscall_routine_open: failed to handle relative path %s for process %s (pid %p), errno %d\n", path,
                proc->name, proc->pid, res);
@@ -27,7 +27,7 @@ ssize_t syscall_routine_open(const char *restrict path, syscall_open_flags_t fla
         return res;
     }
 
-    ssize_t fd = proc_alloc_fd(proc, handle);
+    int fd = proc_alloc_fd(proc, handle);
     if (fd == -1) {
         printf("syscall_routine_open: no available fd for process %s (pid %p)\n", proc->name, proc->pid);
         vfs_close(handle);
@@ -37,7 +37,7 @@ ssize_t syscall_routine_open(const char *restrict path, syscall_open_flags_t fla
     return fd;
 }
 
-int syscall_routine_close(size_t fd, proc_t *proc) {
+int syscall_routine_close(int fd, proc_t *proc) {
     vfs_handle_t *handle = proc_get_fd(proc, fd);
     if (!handle) {
         printf("syscall_routine_close: invalid fd %p for process %s (pid %p)\n", fd, proc->name, proc->pid);
