+++ src/syscall/fs/syscall_fs.h
@@ -25,9 +25,9 @@
 
 void syscall_handle_fs_call(registers_t *regs, proc_t *proc, pml4_t *current_pml4);
 
-ssize_t syscall_fs_helper_get_handle_and_check_backing_node(size_t fd, proc_t *proc, vfs_handle_t **out_handle);
-ssize_t syscall_fs_helper_get_dir_handle(size_t fd, proc_t *proc, vfs_handle_t **out_handle);
-ssize_t syscall_fs_helper_get_not_dir_handle(size_t fd, proc_t *proc, vfs_handle_t **out_handle);
+int syscall_fs_helper_get_handle_and_check_backing_node(int fd, proc_t *proc, vfs_handle_t **out_handle);
+int syscall_fs_helper_get_dir_handle(int fd, proc_t *proc, vfs_handle_t **out_handle);
+int syscall_fs_helper_get_not_dir_handle(int fd, proc_t *proc, vfs_handle_t **out_handle);
 
 /**
  * Basic architecture for io-related syscalls:
@@ -46,31 +46,32 @@ ssize_t syscall_fs_helper_get_not_dir_handle(size_t fd, proc_t *proc, vfs_handle
 
 // count must be <= SSIZE_MAX
 // These two functions don't work on directories; use getdents in place of read for directories.
-ssize_t syscall_routine_read(size_t fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4);
-ssize_t syscall_routine_write(size_t fd, const void *buf, size_t count, proc_t *proc);
+ssize_t syscall_routine_read(int fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4);
+ssize_t syscall_routine_write(int fd, const void *buf, size_t count, proc_t *proc);
 
 // If mode_if_create is used, the only bits that are honored are the permission bits (lower 9) and the special bits
 // (next 3 lowest: setuid, setgid, sticky). The create and directory flags can't be specified simultaneously; if the
 // create flag is set, only regular files can be created.
-ssize_t syscall_routine_open(const char *restrict path, syscall_open_flags_t flags, mode_t mode, proc_t *proc);
-int syscall_routine_close(size_t fd, proc_t *proc);
+int syscall_routine_open(const char *restrict path, syscall_open_flags_t flags, mode_t mode, proc_t *proc);
+int syscall_routine_close(int fd, proc_t *proc);
 
 // These functions fill out_stat, which must be a user-space pointer to a struct kstat
-ssize_t syscall_routine_stat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
-ssize_t syscall_routine_fstat(size_t fd, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
-ssize_t syscall_routine_lstat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
+int syscall_routine_stat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
+int syscall_routine_fstat(int fd, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
+int syscall_routine_lstat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4);
 
 // TODO: implement
 void syscall_routine_poll();
 
 // Unlike read and write, lseek also works on directories
-ssize_t syscall_routine_lseek(size_t fd, ssize_t offset, int whence, proc_t *proc);
+off_t syscall_routine_lseek(int fd, off_t offset, int whence, proc_t *proc);
 
 // This works only on directory file descriptors; count = size of output buffer, NOT number of entries to read
-ssize_t syscall_routine_getdents(size_t fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4);
+// This syscall is the equivalent of Linux's getdents64, NOT getdents.
+ssize_t syscall_routine_getdents(int fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4);
 
-ssize_t syscall_routine_mkdir(const char *restrict path, mode_t mode, proc_t *proc);
-ssize_t syscall_routine_rmdir(const char *restrict path, proc_t *proc);
+int syscall_routine_mkdir(const char *restrict path, mode_t mode, proc_t *proc);
+int syscall_routine_rmdir(const char *restrict path, proc_t *proc);
 void syscall_routine_rename();
 void syscall_routine_chmod();
 void syscall_routine_fchmod();
@@ -78,8 +79,8 @@ void syscall_routine_chown();
 void syscall_routine_fchown();
 void syscall_routine_lchown();
 // TODO: modify file groups
-ssize_t syscall_routine_getcwd(uint64_t out_buf_ptr, size_t size, proc_t *proc, pml4_t *current_pml4);
-ssize_t syscall_routine_chdir(const char *restrict path, proc_t *proc);
+ssize_t syscall_routine_getcwd(uint64_t out_buf_ptr, ssize_t size, proc_t *proc, pml4_t *current_pml4);
+int syscall_routine_chdir(const char *restrict path, proc_t *proc);
 void syscall_routine_fchdir();
 // Hard links CANNOT point to directories
 void syscall_routine_link();
